+++ src/index.ts
@@ -36,6 +36,7 @@ export * from "./io/integrations.js";
 export * from "./io/keybindings.js";
 export * from "./io/render.js";
 export * from "./io/safe-path.js";
+export * from "./interfaces/url-fetcher.js";
 export * from "./input.js";
 export * from "./models/catalog.js";
 export * from "./models/settings.js";